Feat: Added HMS Smoke Time filter API endpoint + tests
This is a route that returns smokes that will be ongoing during any part of a time range given. This is according to the current day's data only and the latest observable data. This expects UTC time and can be converted to expect any range of days, but will have to write a function to retrieve the last observation per day.