Load_ces4 data script updated
Updated to return a list of tract objects to provide a more modular use of the functions. The load ces4 manage.py command has been updated to not delete data but to persist data, in case of a change in the data retrieval pipeline. Also, all model parameters were checked to confirm all float fields weren't integers (storage complexity).